Do not persist the synchronization imap session between syncs.
Make the Imap structure / connection a loop local struct and pass it
to functions that need it. Simplifies imap error handling, since the
connection is short lived. With the IDLE loop now only returning when
things have actually changed, we spend much less time syncing so
connections are more likely to time out on us when we're not using them.